requirements.txt
asgiref==3.3.1; python_version >= "3.6"
django==3.1.7; python_version >= "3.6"
djangorestframework==3.12.2; python_version >= "3.5"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pytz==2021.1; python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"